README todo & parse token & middleware

+++ internal/repository/auth_postgres.go
@@ -15,8 +15,8 @@ func NewAuthPostgres(db *sqlx.DB) *AuthPostgres {
 	return &AuthPostgres{db: db}
 }
 
-func (r *AuthPostgres) CreateUser(user todo.User) (int, error) {
-	var id int
+func (r *AuthPostgres) CreateUser(user todo.User) (int64, error) {
+	var id int64
 
 	query := fmt.Sprintf(
 		"INSERT INTO %s (name, username, password_hash) "+
