+++ internal/repository/repository.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Authorization interface {
-	CreateUser(user todo.User) (int, error)
+	CreateUser(user todo.User) (int64, error)
 	GetUser(username, password string) (todo.User, error)
 }
